Force catalog ID in metastore client config to be a string
YAML determines that a field is a number if it looks like one.  And Viper
parses the config file as YAML, then applies the desired structure to it.
So if you specify a catalog ID with a leading zero:

1. YAML sees a number and parses it.
   * The number has a leading zero, which gets dropped.
1. Viper translates the number to a string to place in CatalogID.
1. We just lost the leading zero!

Instead force it to be a string and fail parsing if we got anything else.

Fixes #2996.
